Create README.md
This code is the model of the article (BGMSDDA) researched by Guobo Xie, Jianming Li, Guosheng Gu, Yuping Sun, Zhiyi Lin, Yinting Zhu and Weiming Wang, which is not yet published and is only used for reviewers' needs. Any other download or citation without the consent of all authors will be considered as illegal citation or plagiarism.